projectflow-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx==0.25.2
celery==5.3.4
redis==5.0.1
pydantic-settings==2.1.0
pgvector==0.2.4

# WebSocket and real-time
websockets==12.0
sse-starlette==1.8.2

# Auth providers
authlib==1.3.0
python-oauth2==1.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
isort==5.13.0
mypy==1.7.1